backend/requirements.txt
# Django & REST Framework
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-storages>=1.14.0

# Database & ORM
psycopg2-binary

# Vector Database
qdrant-client>=1.9.0

# AI & Machine Learning
sentence-transformers>=2.7.0
transformers>=4.43
torch>=2.2.2
openai>=1.35.0

# Document Processing
pypdf>=4.2.0
PyPDF2==3.0.1

# Text Processing & Utilities
regex>=2024.5.15
tqdm>=4.66.4
numpy>=2.2.0,<2.3.0

# LangChain (for advanced RAG features)
langchain>=0.3.8,<0.4.0
langchain-core>=0.3.68,<0.4.0
langchain-community>=0.3.8,<0.4.0
langchain-openai>=0.1.9

# Authentication & Security
PyJWT
python-dotenv

# HTTP & API
requests
uvicorn

# AWS & Cloud Storage
boto3>=1.34.0
botocore>=1.34.0

# Computer Vision (for receipt processing)
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88

# Legacy Support (if needed)
langchain-google-genai==2.1.8
langchain-huggingface>=0.0.5
langchain-upstage==0.7.3

# excel
openpyxl